+++ packages/Core/lib/query.d.ts
@@ -54,6 +54,7 @@ export declare class _nanoSQLQuery implements InanoSQLQueryExec {
     _combineRows(rData: any): {};
     _streamAS(row: any): any;
     _orderByRows(a: any, b: any): number;
+    _tableID(): string;
     _createTable(table: InanoSQLTableConfig, alterTable: boolean, complete: () => void, error: (err: any) => void): void;
     _dropTable(table: string, complete: () => void, error: (err: any) => void): void;
     _onError(err: any): void;